+++ lib/common.sh
@@ -29,14 +29,6 @@ is_sbt_native_packager() {
 	fi
 }
 
-has_old_preset_sbt_opts() {
-	if [[ "${SBT_OPTS:-}" = "-Xmx384m -Xss512k -XX:+UseCompressedOops" ]]; then
-		return 0
-	else
-		return 1
-	fi
-}
-
 # sed -l basically makes sed replace and buffer through stdin to stdout
 # so you get updates while the command runs and dont wait for the end
 # e.g. sbt stage | indent
@@ -50,54 +42,23 @@ output() {
 	esac
 }
 
-install_sbt_extras() {
-	local opt_dir="${1}"
-	local sbt_bin_dir="${2}"
-
-	rm -f "${sbt_bin_dir}"/sbt-launch*.jar #legacy launcher
-	mkdir -p "${sbt_bin_dir}"
-	cp -p "${opt_dir}"/sbt-extras.sh "${sbt_bin_dir}"/sbt-extras
-	cp -p "${opt_dir}"/sbt-wrapper.sh "${sbt_bin_dir}"/sbt
-
-	chmod 0755 "${sbt_bin_dir}"/sbt-extras
-	chmod 0755 "${sbt_bin_dir}"/sbt
-
-	export PATH="${sbt_bin_dir}:${PATH}"
-}
-
 run_sbt() {
-	local home="${1}"
-	local tasks="${2}"
+	local tasks="${1}"
 	local build_log_file=".heroku/sbt-build.log"
-
+	mkdir -p "$(dirname "${build_log_file}")"
 	echo "" >"${build_log_file}"
-
-	export SBT_EXTRAS_OPTS="${SBT_EXTRAS_OPTS:-}"
 
 	output::step "Running: sbt ${tasks}"
 	# shellcheck disable=SC2086  # We want word splitting for tasks
-	if ! SBT_HOME="${home}" sbt ${tasks} | output "${build_log_file}"; then
+	if ! sbt ${tasks} | output "${build_log_file}"; then
 		handle_sbt_errors "${build_log_file}"
 		exit 1
 	fi
 }
 
 write_sbt_dependency_classpath_log() {
-	local home="${1}"
-
 	export SBT_EXTRAS_OPTS="${SBT_EXTRAS_OPTS:-}"
 
 	output::step "Collecting dependency information"
-	SBT_HOME="${home}" sbt "show dependencyClasspath" | grep -o "Attributed\(.*\)" >.heroku/sbt-dependency-classpath.log || true
-}
-
-cache_copy() {
-	rel_dir="${1}"
-	from_dir="${2}"
-	to_dir="${3}"
-	rm -rf "${to_dir:?}/${rel_dir}"
-	if [[ -d "${from_dir}/${rel_dir}" ]]; then
-		mkdir -p "${to_dir}/${rel_dir}"
-		cp -pr "${from_dir}/${rel_dir}"/. "${to_dir}/${rel_dir}"
-	fi
+	sbt "show dependencyClasspath" | grep -o "Attributed\(.*\)" >.heroku/sbt-dependency-classpath.log || true
 }