+++ test/spec/cache_spec.rb
@@ -10,19 +10,14 @@ describe 'Scala buildpack' do
         expect(clean_output(app.output)).to eq(<<~OUTPUT)
           remote: -----> Scala app detected
           remote: -----> Installing Azul Zulu OpenJDK $VERSION
+          remote: -----> Downloading sbt launcher 1.11.7...
+          remote: -----> Setting up sbt launcher...
           remote: -----> Running: sbt compile stage
-          remote: Downloading sbt launcher for 1.11.7:
-          remote:   From  https://repo1.maven.org/maven2/org/scala-sbt/sbt-launch/1.11.7/sbt-launch-1.11.7.jar
-          remote:     To  $BUILD_DIR/.sbt_home/launchers/1.11.7/sbt-launch.jar
-          remote: Downloading sbt launcher 1.11.7 md5 hash:
-          remote:   From  https://repo1.maven.org/maven2/org/scala-sbt/sbt-launch/1.11.7/sbt-launch-1.11.7.jar.md5
-          remote:     To  $BUILD_DIR/.sbt_home/launchers/1.11.7/sbt-launch.jar.md5
-          remote:        $BUILD_DIR/.sbt_home/launchers/1.11.7/sbt-launch.jar: OK
           remote: [info] [launcher] getting org.scala-sbt sbt 1.11.7  (this may take some time)...
           remote: [info] [launcher] getting Scala 2.12.20 (for sbt)...
           remote:        [info] welcome to sbt 1.11.7 (Azul Systems, Inc. Java $VERSION)
-          remote:        [info] loading global plugins from $BUILD_DIR/.sbt_home/plugins
-          remote:        [info] compiling 1 Scala source to $BUILD_DIR/.sbt_home/plugins/target/scala-2.12/sbt-1.0/classes ...
+          remote:        [info] loading global plugins from /tmp/codon/tmp/cache/sbt_global/plugins
+          remote:        [info] compiling 1 Scala source to /tmp/codon/tmp/cache/sbt_global/plugins/target/scala-2.12/sbt-1.0/classes ...
           remote:        [info] Non-compiled module 'compiler-bridge_2.12' for Scala 2.12.20. Compiling...
           remote:        [info]   Compilation completed in $DURATION.
           remote:        [info] done compiling
@@ -48,14 +43,12 @@ describe 'Scala buildpack' do
           remote:        [info] Wrote $BUILD_DIR/target/scala-2.13/sbt-1-11-7-play-3-x-scala-2-13-x_2.13-1.0-SNAPSHOT.pom
           remote:        [success] Total time: $DURATION, completed $DATETIME
           remote: -----> Collecting dependency information
-          remote: -----> Dropping ivy cache from the slug
-          remote: -----> Dropping sbt boot dir from the slug
-          remote: -----> Dropping sbt cache dir from the slug
+          remote: -----> Copying sbt and dependencies into slug for runtime use
           remote: -----> Dropping compilation artifacts from the slug
           remote: -----> Discovering process types
           remote:        Procfile declares types     -> (none)
           remote:        Default types for buildpack -> web
-          
+
           remote: -----> Compressing...
           remote:        Done: 114M
         OUTPUT
@@ -67,9 +60,10 @@ describe 'Scala buildpack' do
         expect(clean_output(app.output)).to eq(<<~OUTPUT)
           remote: -----> Scala app detected
           remote: -----> Installing Azul Zulu OpenJDK $VERSION
+          remote: -----> Setting up sbt launcher...
           remote: -----> Running: sbt compile stage
           remote:        [info] welcome to sbt 1.11.7 (Azul Systems, Inc. Java $VERSION)
-          remote:        [info] loading global plugins from $BUILD_DIR/.sbt_home/plugins
+          remote:        [info] loading global plugins from /tmp/codon/tmp/cache/sbt_global/plugins
           remote:        [info] loading settings for project $BUILD_ID-build from plugins.sbt...
           remote:        [info] loading project definition from $BUILD_DIR/project
           remote:        [info] loading settings for project root from build.sbt...
@@ -90,14 +84,12 @@ describe 'Scala buildpack' do
           remote:        [info] Wrote $BUILD_DIR/target/scala-2.13/sbt-1-11-7-play-3-x-scala-2-13-x_2.13-1.0-SNAPSHOT.pom
           remote:        [success] Total time: $DURATION, completed $DATETIME
           remote: -----> Collecting dependency information
-          remote: -----> Dropping ivy cache from the slug
-          remote: -----> Dropping sbt boot dir from the slug
-          remote: -----> Dropping sbt cache dir from the slug
+          remote: -----> Copying sbt and dependencies into slug for runtime use
           remote: -----> Dropping compilation artifacts from the slug
           remote: -----> Discovering process types
           remote:        Procfile declares types     -> (none)
           remote:        Default types for buildpack -> web
-          
+
           remote: -----> Compressing...
           remote:        Done: 114M
         OUTPUT
@@ -122,7 +114,7 @@ describe 'Scala buildpack' do
         expect(clean_output(app.output)).to eq(<<~OUTPUT)
           remote: -----> Scala app detected
           remote: -----> Installing Azul Zulu OpenJDK $VERSION
-          
+
           remote:  !     Warning: Unsupported sbt version detected.
           remote:  !
           remote:  !     This buildpack does not officially support sbt 0.13.18. You are using
@@ -139,15 +131,13 @@ describe 'Scala buildpack' do
           remote:  !
           remote:  !     Upgrade guide:
           remote:  !     - https://www.scala-sbt.org/1.x/docs/Migrating-from-sbt-013x.html
-          
+
+          remote: -----> Downloading sbt launcher 0.13.18...
+          remote: -----> Setting up sbt launcher...
           remote: -----> Running: sbt compile stage
-          remote: Downloading sbt launcher for 0.13.18:
-          remote:   From  https://repo.typesafe.com/typesafe/ivy-releases/org.scala-sbt/sbt-launch/0.13.18/sbt-launch.jar
-          remote:     To  $BUILD_DIR/.sbt_home/launchers/0.13.18/sbt-launch.jar
-          remote:        
           remote:        Getting org.scala-sbt sbt 0.13.18  (this may take some time)...
           remote: WARNING: An illegal reflective access operation has occurred
-          remote: WARNING: Illegal reflective access by org.apache.ivy.util.url.IvyAuthenticator (file:$BUILD_DIR/.sbt_home/launchers/0.13.18/sbt-launch.jar) to field java.net.Authenticator.theAuthenticator
+          remote: WARNING: Illegal reflective access by org.apache.ivy.util.url.IvyAuthenticator (file:/tmp/codon/tmp/cache/sbt-launcher/sbt-launch-0.13.18.jar) to field java.net.Authenticator.theAuthenticator
           remote: WARNING: Please consider reporting this to the maintainers of org.apache.ivy.util.url.IvyAuthenticator
           remote: WARNING: Use --illegal-access=warn to enable warnings of further illegal reflective access operations
           remote: WARNING: All illegal access operations will be denied in a future release
@@ -264,8 +254,8 @@ describe 'Scala buildpack' do
           remote:        :: retrieving :: org.scala-sbt#boot-scala
           remote:        	confs: [default]
           remote:        	5 artifacts copied, 0 already retrieved (24560kB/$DURATION)
-          remote:        [info] Loading global plugins from $BUILD_DIR/.sbt_home/plugins
-          remote:        [info] Updating {file:$BUILD_DIR/.sbt_home/plugins/}global-plugins...
+          remote:        [info] Loading global plugins from /tmp/codon/tmp/cache/sbt_global/plugins
+          remote:        [info] Updating {file:/tmp/codon/tmp/cache/sbt_global/plugins/}global-plugins...
           remote:        [info] Resolving org.scala-lang#scala-library;2.10.7 ...
           remote:        [info] Resolving org.scala-sbt#sbt;0.13.18 ...
           remote:        [info] Resolving org.scala-sbt#main;0.13.18 ...
@@ -320,7 +310,7 @@ describe 'Scala buildpack' do
           remote:        [info] Resolving org.scala-lang#jline;2.10.7 ...
           remote:        [info] Resolving org.fusesource.jansi#jansi;1.4 ...
           remote:        [info] Done updating.
-          remote:        [info] Compiling 1 Scala source to $BUILD_DIR/.sbt_home/plugins/target/scala-2.10/sbt-0.13/classes...
+          remote:        [info] Compiling 1 Scala source to /tmp/codon/tmp/cache/sbt_global/plugins/target/scala-2.10/sbt-0.13/classes...
           remote:        [info] 'compiler-interface' not yet compiled for Scala 2.10.7. Compiling...
           remote:        [info]   Compilation completed in $DURATION.
           remote:        [info] Loading project definition from $BUILD_DIR/project
@@ -837,17 +827,16 @@ describe 'Scala buildpack' do
           remote:        [success] Total time: $DURATION, completed $DATETIME
           remote: -----> Collecting dependency information
           remote: WARNING: An illegal reflective access operation has occurred
-          remote: WARNING: Illegal reflective access by sbt.ivyint.ErrorMessageAuthenticator$ (file:$BUILD_DIR/.sbt_home/boot/scala-2.10.7/org.scala-sbt/sbt/0.13.18/ivy-0.13.18.jar) to field java.net.Authenticator.theAuthenticator
+          remote: WARNING: Illegal reflective access by sbt.ivyint.ErrorMessageAuthenticator$ (file:/tmp/codon/tmp/cache/sbt_boot/scala-2.10.7/org.scala-sbt/sbt/0.13.18/ivy-0.13.18.jar) to field java.net.Authenticator.theAuthenticator
           remote: WARNING: Please consider reporting this to the maintainers of sbt.ivyint.ErrorMessageAuthenticator$
           remote: WARNING: Use --illegal-access=warn to enable warnings of further illegal reflective access operations
           remote: WARNING: All illegal access operations will be denied in a future release
-          remote: -----> Dropping ivy cache from the slug
-          remote: -----> Dropping sbt boot dir from the slug
+          remote: -----> Copying sbt and dependencies into slug for runtime use
           remote: -----> Dropping compilation artifacts from the slug
           remote: -----> Discovering process types
           remote:        Procfile declares types     -> (none)
           remote:        Default types for buildpack -> web
-          
+
           remote: -----> Compressing...
           remote:        Done: 99.8M
         OUTPUT
@@ -876,12 +865,13 @@ describe 'Scala buildpack' do
           remote:  !
           remote:  !     Upgrade guide:
           remote:  !     - https://www.scala-sbt.org/1.x/docs/Migrating-from-sbt-013x.html
-          
+
+          remote: -----> Setting up sbt launcher...
           remote: -----> Running: sbt compile stage
-          remote:        [info] Loading global plugins from $BUILD_DIR/.sbt_home/plugins
-          remote:        [info] Updating {file:$BUILD_DIR/.sbt_home/plugins/}global-plugins...
+          remote:        [info] Loading global plugins from /tmp/codon/tmp/cache/sbt_global/plugins
+          remote:        [info] Updating {file:/tmp/codon/tmp/cache/sbt_global/plugins/}global-plugins...
           remote: WARNING: An illegal reflective access operation has occurred
-          remote: WARNING: Illegal reflective access by sbt.ivyint.ErrorMessageAuthenticator$ (file:$BUILD_DIR/.sbt_home/boot/scala-2.10.7/org.scala-sbt/sbt/0.13.18/ivy-0.13.18.jar) to field java.net.Authenticator.theAuthenticator
+          remote: WARNING: Illegal reflective access by sbt.ivyint.ErrorMessageAuthenticator$ (file:/tmp/codon/tmp/cache/sbt_boot/scala-2.10.7/org.scala-sbt/sbt/0.13.18/ivy-0.13.18.jar) to field java.net.Authenticator.theAuthenticator
           remote: WARNING: Please consider reporting this to the maintainers of sbt.ivyint.ErrorMessageAuthenticator$
           remote: WARNING: Use --illegal-access=warn to enable warnings of further illegal reflective access operations
           remote: WARNING: All illegal access operations will be denied in a future release
@@ -939,7 +929,7 @@ describe 'Scala buildpack' do
           remote:        [info] Resolving org.scala-lang#jline;2.10.7 ...
           remote:        [info] Resolving org.fusesource.jansi#jansi;1.4 ...
           remote:        [info] Done updating.
-          remote:        [info] Compiling 1 Scala source to $BUILD_DIR/.sbt_home/plugins/target/scala-2.10/sbt-0.13/classes...
+          remote:        [info] Compiling 1 Scala source to /tmp/codon/tmp/cache/sbt_global/plugins/target/scala-2.10/sbt-0.13/classes...
           remote:        [info] Loading project definition from $BUILD_DIR/project
           remote:        [info] Updating {file:$BUILD_DIR/project/}$BUILD_ID-build...
           remote:        [info] Resolving org.scala-sbt#global-plugins;0.0 ...
@@ -1113,19 +1103,18 @@ describe 'Scala buildpack' do
           remote:        [success] Total time: $DURATION, completed $DATETIME
           remote: -----> Collecting dependency information
           remote: WARNING: An illegal reflective access operation has occurred
-          remote: WARNING: Illegal reflective access by sbt.ivyint.ErrorMessageAuthenticator$ (file:$BUILD_DIR/.sbt_home/boot/scala-2.10.7/org.scala-sbt/sbt/0.13.18/ivy-0.13.18.jar) to field java.net.Authenticator.theAuthenticator
+          remote: WARNING: Illegal reflective access by sbt.ivyint.ErrorMessageAuthenticator$ (file:/tmp/codon/tmp/cache/sbt_boot/scala-2.10.7/org.scala-sbt/sbt/0.13.18/ivy-0.13.18.jar) to field java.net.Authenticator.theAuthenticator
           remote: WARNING: Please consider reporting this to the maintainers of sbt.ivyint.ErrorMessageAuthenticator$
           remote: WARNING: Use --illegal-access=warn to enable warnings of further illegal reflective access operations
           remote: WARNING: All illegal access operations will be denied in a future release
-          remote: -----> Dropping ivy cache from the slug
-          remote: -----> Dropping sbt boot dir from the slug
+          remote: -----> Copying sbt and dependencies into slug for runtime use
           remote: -----> Dropping compilation artifacts from the slug
           remote: -----> Discovering process types
           remote:        Procfile declares types     -> (none)
           remote:        Default types for buildpack -> web
-          
+
           remote: -----> Compressing...
-          remote:        Done: 99.9M
+          remote:        Done: 99.8M
         OUTPUT
       end
     end